osd: Disable heartbeat timeout until a non-future workitem can be processed
There could be rare instances when emplying mclock scheduler where a worker
thread for a shard may not get an immediate work item to process. Such
items are designated as future work items. In such cases, the _process()
loop waits until the time indicated by the scheduler to attempt a dequeue
from the scheduler queue again. It may so happen that if there are multiple
threads per shard, a thread may not get an immediate item for a long time.
This time could exceed the heartbeat timeout for the thread and result in
hearbeat timeouts reported for the osd in question. To prevent this, the
heartbeat timeouts for the thread is disabled before waiting and enabled
once the wait period is over.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>