mon/LogMonitor: replace semi-broken pending_summary with pending_keys
First, we don't need to store actual keys in pending_summary.  That is
what pending_log contains (the actual pending entries).

However, we do want to efficiently track pending_keys so that we ignore
dups within the same proposal interval.  Use pending_keys for that.

Fix a few users to look only and summary.keys or both summary and
pending_keys, as appropriate.

This is simpler, less confusing, and probably fixes some very subtle
bugs.

Signed-off-by: Sage Weil <sage@newdream.net>