mgr/volumes: Fail subvolume removal if it's not completed/cancelled
Removing subvolume clone which is not either complete or cancelled
with force doesn't remove the clone index (tracker). This results
in the cloner thread to stuck in loop trying to clone the deleted
one.

This patch addresses the issue by not allowing the subvolume clone
to be removed if it's not complete/cancelled even with force option.
It throws the error EAGAIN, asking the user to cancel the pending clone
and retry.

Fixes: https://tracker.ceph.com/issues/51707
Signed-off-by: Kotresh HR <khiremat@redhat.com>