os/bluestore: simplify zone to clean selection
Only pick one zone to clean based on the current.  Since the best victim
may change (maybe another zone gets a bunch of releases and new dead
bytes!) there is no reason (yet) to explicitly avoid the victim zone
during allocation.  There is also no need to track which zones we are
cleaning on disk because we can choose to clean from any zone at any time,
and in general want to clean from the best candidate at the time, not the
one that looked the best some time in the past.

Signed-off-by: Sage Weil <sage@newdream.net>